Add async endpoints, tests, and coverage support (#4)
Migrates controller endpoints to async, adds unit/integration tests for API endpoints, and introduces code coverage support via CMake and helper script. Updates Dockerfile to install lcov, improves .gitignore, and enhances README with testing and coverage instructions.

+++ src/worker/CpuMock.cpp
@@ -6,7 +6,7 @@
 namespace app { namespace worker {
 
 void AudioWorker::computeMelSpectrogram(const std::vector<float>& inputAudio, std::vector<float>& outputMel) {    
-    OATPP_LOGD("AudioWorker", "[MOCK-CPU] Receiving Audio Data Size: %d", inputAudio.size());
+    OATPP_LOGD("AudioWorker", "[MOCK-CPU] Receiving Audio Data Size: %ld", (long)inputAudio.size());
     OATPP_LOGD("AudioWorker", "[MOCK-CPU] Simulating CUDA processing...");
     
     std::this_thread::sleep_for(std::chrono::milliseconds(500));
